[LTO] Override TargetABI from module flags if present when creating TargetMachine
RISC-V's data layout is also determined by the ABI, not only the target triple,
however the TargetMachine is created with the target triple's data layout,
that is not always correct. This patch changes the TargetMachine's data
layout to the one specified in the module flags if present.

The same problem will happen with other targets like MIPS, but unfortunately,
MIPS didn't emit the target-abi into the module flags, so this patch only
fixes the issue for RISC-V.

NOTE: MIPS with -mabi=n32 can trigger the same issue.

Another possible solution is add new parameter to the TargetMachine
constructor, but that would require changes in all the targets.